% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{BioUtils-package}
\alias{BioUtils}
\alias{BioUtils-package}
\title{BioUtils: Tools for Biological Data Analysis and Visualization}
\description{
BioUtils provides a collection of tools for preprocessing biological
datasets, performing statistical analyses, and generating visualizations.
See \code{vignette("bioutils-case-study")} for a full workflow walkthrough.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spencertreadway.github.io/BioUtils/}
  \item \url{https://github.com/spencertreadway/BioUtils}
}

}
\author{
\strong{Maintainer}: Spencer Treadway \email{spencer.treadway08@gmail.com}

Authors:
\itemize{
  \item Spencer Treadway \email{spencer.treadway08@gmail.com}
}

}
\keyword{internal}
