% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioutils.R
\docType{package}
\name{bioutils}
\alias{bioutils}
\alias{bioutils-package}
\title{bioutils: Useful functions for computational biology}
\description{
Various functions useful for computational biology.
}
\section{bioutils functions}{

    \code{ensembl2hgnc}
    \code{rsid2varinfo}
}

