% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTtoRE.R
\name{CTtoRE}
\alias{CTtoRE}
\title{CT to RE}
\usage{
CTtoRE(CTvals, ctrl, hskp)
}
\arguments{
\item{CTvals}{the ct values under examination from the RQdeltaCT package}

\item{ctrl}{string defining the control group in the CTvals tibble (Group col)}

\item{hskp}{string defining the housekeeping gene in the CTvals tibble (which Gene col)}
}
\value{
a tibble containing 5 rows with the calculated data
}
\description{
A simple script that takes in a tibble of CT values as produced by the RQdeltaCT package and
calculates the deltaCT, 2^-(delta delta CT), and relative expression values and returns them as
separate tibbles in a tibble that also contains lists of the mean for the control group CT, and
the mean of the control group for 2^-(delta delta CT) values. In addition to the CT values the
name of the control group and the desired housekeeping gene name must be supplied.

Returned Tibble:

CT_ctrl_mean: a list of the mean CT values for the control group of each gene

dCT: a tibble of the deltaCT values

ddCT: a tibble of the 2^-(delta delta CT) values

ddCT_ctrl_mean: a list of the mean ddCT values for the control group of each gene

RE: relative expression for each sample as the fold change compared to the control group

Currently NO error checking of the inputs is being carried out!
}
