% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl2hgnc.R
\name{ensembl2hgnc}
\alias{ensembl2hgnc}
\title{Ensembl ids 2 hgnc}
\usage{
ensembl2hgnc(ensembl_gene_ids, host = "grch37.ensembl.org",
  drop_dot_ensembl_id = TRUE)
}
\arguments{
\item{ensembl_gene_ids}{Character vector.
List of Ensembl gene ids to get hgnc symbols for.}

\item{host}{Character.
Ensembl biomaRt host.}

\item{drop_dot_ensembl_id}{Logical.
Drop "." from ENSG00000072310.12}
}
\value{
Character.
    The corresponding hgnc symbols to the Ensembl ids.
}
\description{
\code{ensembl2hgnc} converts Ensembl gene ids to hgnc symbols. If no
hgnc symbol then uses external_gene_name. If no external_gene_name then
uses Ensembl gene id.
}
\examples{
ensembl2hgnc(c("ENSG00000109501", "ENSG00000058453", "ENSG00000030066"))

}
