% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_query.R
\name{find.probe.by.gene}
\alias{find.probe.by.gene}
\title{Find Probe ID by Gene Name}
\usage{
find.probe.by.gene(genes, gene.name)
}
\arguments{
\item{genes}{Data frame of gene annotations.}

\item{gene.name}{Character. Target gene name.}
}
\value{
Character. Probe ID.
}
\description{
Searches feature data to find the probe ID corresponding
to a given gene name or description.
}
