% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid2varinfo.R
\name{rsid2varinfo}
\alias{rsid2varinfo}
\title{rsid to variant info.}
\usage{
rsid2varinfo(rs_ids, host = "grch37.ensembl.org",
  var_attributes = c("refsnp_id", "chr_name", "chrom_start", "allele"))
}
\arguments{
\item{rs_ids}{Character vector.
List of rs gene ids to get information for.}

\item{host}{Character.
Ensembl biomaRt host.}

\item{var_attributes}{Character vector.
List of attributes to select from Ensembl.}
}
\value{
Data.frame.
    Data.frame of the attributes requested. Note that duplicates are removed
    and the order is not the same as given.
}
\description{
\code{rsid2varinfo} looks up information from var_attributes for rs ids.
}
\examples{
rsid2varinfo(c("rs8050136", "rs780094", "rs16942", "rs8050136"))

}
